Added custom pendulum env to work with TD3
Built upon TD3_basic, this notebook has additional step: after we get a trained policy on a basic pendulum environment (same as in TD3_basic.ipynb), we then initialize new custom pendulum anvironemnt and try to train a new policy using the reward function we estimate from the old policy. Currently the code run, but the model does not successfully train on the estimation of the reward function. Things changed in the custom env from Tim's version: freward transferred to cuda, model actions adjusted for TD3 model, observations progress across the timesteps when collected by the funciton for estimation of the reward.